--- src/components/Button.css
@@ -1,19 +1,0 @@
-.btn {
-	padding: 8px;
-	box-shadow: 0 0 40px 40px #3498db inset, 0 0 0 0 #3498db;
-	transition: all 150ms ease-in-out;
-	text-align: center;
-	width: 50%;
-	margin: 0% auto;
-	padding: 8px;
-	text-decoration: none;
-	color: #fff;
-	font-weight: 600;
-	margin-bottom: 20px;
-}
-
-.btn:hover {
-	color: #000;
-	box-shadow: 0 0 10px 0 #3498db inset, 0 0 10px 4px #3498db;
-	cursor: pointer;
-}
