--- src/components/GoHomeBtn.css
@@ -1,23 +1,0 @@
-.go-home-btn {
-	position: fixed;
-	bottom: 30px;
-	right: 1rem;
-	width: 40px;
-	height: 37px;
-	box-shadow: 0px 4px 31px rgba(0 0 0 0.15);
-	border-style: none;
-	background: #3498db;
-	z-index: 10;
-	border-radius: 4px;
-}
-
-.go-home-btn:hover {
-	cursor: pointer;
-	box-shadow: 1px 10px 32px rgba(0 0 0 0.45);
-	transition: all 0.2s ease-in-out;
-	background: #0e5e94;
-}
-
-.home-icon {
-	width: 25px;
-}
