+++ src/images.css
@@ -1,0 +1,72 @@
+.main {
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	width: 100%;
+	height: 100%;
+	background-color: #f0f0f0;
+	margin: auto;
+	flex-direction: column;
+}
+
+.container {
+	display: flex;
+	flex-wrap: wrap;
+	justify-content: center;
+	object-fit: cover;
+	object-position: top center;
+	width: 100%;
+	height: 100%;
+	max-width: 100%;
+	max-height: 100%;
+	padding: 50px;
+}
+
+.imageContainer {
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	align-items: center;
+	align-content: center;
+	cursor: pointer;
+	width: 300px;
+	min-height: 300px;
+	border-radius: 5px;
+	overflow: hidden;
+	position: relative;
+	transition: transform 0.3s;
+}
+
+.imageWrapper {
+	width: 100%;
+	padding-top: 100%;
+	position: relative;
+}
+
+.image {
+	width: 100%;
+	height: 100%;
+	object-fit: cover;
+	object-position: top center;
+	position: absolute;
+	top: 0;
+	left: 0;
+}
+
+.imageDescription {
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	align-content: center;
+	width: 100%;
+	height: 50px;
+	background-color: #000000;
+	color: #d1d1d1;
+	border-radius: 0 0 5px 5px;
+}
+
+.imageDescriptionText {
+	font-size: 16px;
+	font-weight: bold;
+	margin: 0;
+}
